Raise an exception in case of deprecated IS0 639-3
When Lang is initiated with a deprecated ISO 639-3 code, a
DeprecatedLanguageValue exception is raised.

The name of the deprecated language is returned in the 'name' attribute
of the exception. When there is one, the new ISO 639-3 code is returned
in the 'change_to' attribute of the exception.

All necessary data is loaded from the official local file named
'iso-639-3_Retirements.tab'.